requirements.txt
absl-py
python-dotenv
altair
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
astunparse
async-lru
attrs
babel
beautifulsoup4
bleach
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
comm
contourpy
cycler
debugpy
decorator
defusedxml
executing
fastjsonschema
filelock
flatbuffers
fonttools
fqdn
fsspec
gast
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google-pasta
googleapis-common-protos
grpcio
grpcio-status
h11
h5py
httpcore
httplib2
httpx
idna
ipykernel
ipython
isoduration
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter-server-mathjax
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_git
jupyterlab_pygments
jupyterlab_server
keras
kiwisolver
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
ml-dtypes
mpmath
namex
narwhals
nbclient
nbconvert
nbdime
nbformat
nest-asyncio
networkx
notebook_shim
numpy
opt-einsum
optree
opencv-python
overrides
packaging
pandas
pandocfilters
parso
pexpect
pillow
platformdirs
plotly
prometheus_client
prompt_toolkit
proto-plus
protobuf
psutil
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydeck
Pygments
pyparsing
python-dateutil
python-json-logger
pytz
PyYAML
pyzmq
referencing
requests
rfc3339-validator
rfc3986-validator
rich
rpds-py
rsa
scikit-learn
scipy
seaborn
Send2Trash
setuptools
six
smmap
sniffio
soupsieve
stack-data
streamlit
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
termcolor
terminado
threadpoolctl
tinycss2
toml
torch
tornado
tqdm
traitlets
types-python-dateutil
typing_extensions
tzdata
uri-template
uritemplate
urllib3
watchdog
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
wheel
wrapt